Separate CreateScenarioButton and CreateScenarioDialog, passing props to HierarchicalComboBox Autocomplete
