+++ src/components/CreateScenarioDialog/CreateScenarioButton.js
@@ -1,0 +1,57 @@
+import React, { useState } from 'react';
+import PropTypes from 'prop-types';
+import { withStyles } from '@material-ui/core/styles';
+import {
+  Button, Typography
+} from '@material-ui/core';
+import AddIcon from '@material-ui/icons/Add';
+import { useTranslation } from 'react-i18next';
+import CreateScenarioDialog from './CreateScenarioDialog';
+
+const useStyles = theme => ({
+  root: {
+    height: '100%'
+  },
+  dialogContent: {
+    marginLeft: '20px',
+    marginRight: '20px'
+  },
+  dialogActions: {
+    marginRight: '30px',
+    marginTop: '20px',
+    marginBottom: '5px'
+  }
+});
+
+const CreateScenarioButton = ({ classes, currentScenario, datasets, scenarios, runTemplates }) => {
+  const { t } = useTranslation();
+  const [open, setOpen] = useState(false);
+  const openDialog = () => setOpen(true);
+  const closeDialog = () => setOpen(false);
+
+  return (
+    <div>
+      <Button size="medium" startIcon={<AddIcon />} variant="text" onClick={openDialog} color="primary">
+        <Typography noWrap color="primary">{t('commoncomponents.button.create.scenario', 'Create Alternative Scenario')}</Typography>
+      </Button>
+      <CreateScenarioDialog
+          open={open}
+          currentScenario={currentScenario}
+          datasets={datasets}
+          classes={classes}
+          closeDialog={closeDialog}
+          runTemplates={runTemplates}
+          scenarios={scenarios} />
+    </div>
+  );
+};
+
+CreateScenarioButton.propTypes = {
+  classes: PropTypes.any,
+  currentScenario: PropTypes.object,
+  scenarios: PropTypes.array.isRequired,
+  datasets: PropTypes.array.isRequired,
+  runTemplates: PropTypes.array.isRequired
+};
+
+export default withStyles(useStyles)(CreateScenarioButton);
