+++ src/components/CreateScenarioDialog/index.js
@@ -1,0 +1,5 @@
+// Copyright (c) Cosmo Tech.
+// Licensed under the MIT license.
+
+export { default as CreateScenarioButton } from './CreateScenarioButton';
+export { default as CreateScenarioDialog } from './CreateScenarioDialog';
